Modificacion de container y componentes del alumno. Separacion de logica y renderizacion
