+++ src/components/Alumno/AgregarAlumno.jsx
@@ -4,17 +4,21 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
 import { useState } from 'react';
 //Components
-import InputComponent from '../InputComponent';
-import FeedbackText from '../FeedbackText';
+import InputComponent from '../Utils/InputComponent';
+import InputReComponent from '../Utils/InputReComponent';
 import NacimientoComponent from '../Utils/NacimientoComponent';
 
 const AgregarAlumno = ({
   active,
   setActive,
-  setAlumnos,
   alumnos,
   setActAlumnos,
   setAlumnosLoader,
+  handleChangeName,
+  handleChangePhone,
+  handleSubmitAlumnoForm,
+  setAlumnoForm,
+  alumnoForm
 }) => {
   const URL_BASE = `http://localhost:8083/api/`;
   const [nombre, setNombre] = useState('');
@@ -26,107 +30,101 @@ const AgregarAlumno = ({
   const [telefonoFB, setTelefonoFB] = useState({ text: '', color: '' });
 
   //estos para la fecha de nacimiento personalizada
-  const handleChangeName = (e) => {
-    const pattern = new RegExp('^[A-Z]+$', 'i');
-    const word = e.target.value.split(' ').join('');
-    const submitBtn = document.getElementById('alumno-add-form-addBtn');
+  // const handleChangeName = (e) => {
+  //   const pattern = new RegExp('^[A-Z]+$', 'i');
+  //   const word = e.target.value.split(' ').join('');
+  //   const submitBtn = document.getElementById('alumno-add-form-addBtn');
 
-    //validar que el nombrte sea solo texto y que no exista repetidos
-    setNombre(e.target.value);
-    const nextInput = document.getElementById('telefonotinput');
-    if (e.target.value === '') {
-      setNombreFB({ ...nombreFB, text: '', color: '' });
-      nextInput.disabled = true;
-      submitBtn.disabled = true;
-    } else {
-      //Cumple las expectativas de ser un nombre
-      if (pattern.test(word)) {
-        if (
-          alumnos
-            .map((each) => each.nombre.toUpperCase())
-            .indexOf(e.target.value.toUpperCase()) === -1
-        ) {
-          setNombreFB({
-            ...nombreFB,
-            text: 'El nombre de alumno es correcto',
-            color: '#7CBD1E',
-          });
-          nextInput.disabled = false;
-        } else {
-          setNombreFB({
-            ...nombreFB,
-            text: 'El nombre de usuario ya existe',
-            color: '#CC3636',
-          });
-          nextInput.disabled = true;
-          submitBtn.disabled = true;
-        }
-      } else {
-        setNombreFB({
-          ...nombreFB,
-          text: 'Escriba un nombre de usuario sin numeros',
-          color: '#CC3636',
-        });
-      }
-    }
-  };
+  //   //validar que el nombrte sea solo texto y que no exista repetidos
+  //   setNombre(e.target.value);
+  //   const nextInput = document.getElementById('telefonotinput');
+  //   if (e.target.value === '') {
+  //     setNombreFB({ ...nombreFB, text: '', color: '' });
+  //     nextInput.disabled = true;
+  //     submitBtn.disabled = true;
+  //   } else {
+  //     //Cumple las expectativas de ser un nombre
+  //     if (pattern.test(word)) {
+  //       if (
+  //         alumnos
+  //           .map((each) => each.nombre.toUpperCase())
+  //           .indexOf(e.target.value.toUpperCase()) === -1
+  //       ) {
+  //         setNombreFB({
+  //           ...nombreFB,
+  //           text: 'El nombre de alumno es correcto',
+  //           color: '#7CBD1E',
+  //         });
+  //         nextInput.disabled = false;
+  //       } else {
+  //         setNombreFB({
+  //           ...nombreFB,
+  //           text: 'El nombre de usuario ya existe',
+  //           color: '#CC3636',
+  //         });
+  //         nextInput.disabled = true;
+  //         submitBtn.disabled = true;
+  //       }
+  //     } else {
+  //       setNombreFB({
+  //         ...nombreFB,
+  //         text: 'Escriba un nombre de usuario sin numeros',
+  //         color: '#CC3636',
+  //       });
+  //     }
+  //   }
+  // };
 
-  const handleChangePhone = (e) => {
-    const pattern = '^[0-9]+$';
-    const tel = e.target.value;
-    setTelefono(tel);
-    const submitBtn = document.getElementById('alumno-add-form-addBtn');
+  // const handleChangePhone = (e) => {
+  //   const pattern = '^[0-9]+$';
+  //   const tel = e.target.value;
+  //   setTelefono(tel);
+  //   const submitBtn = document.getElementById('alumno-add-form-addBtn');
 
-    if (tel === '') {
-      setTelefonoFB({ ...telefonoFB, text: '', color: '' });
-      submitBtn.disabled = true;
-    } else {
-      if (tel.match(pattern) != null && telefono.length >= 6) {
-        setTelefonoFB({
-          ...telefonoFB,
-          text: 'El nummero de telefono es correcto',
-          color: '#7CBD1E',
-        });
-        submitBtn.disabled = false;
-      } else {
-        setTelefonoFB({
-          ...telefonoFB,
-          text: 'Solo numeros, minimo 7',
-          color: '#CC3636',
-        });
-        submitBtn.disabled = true;
-      }
-    }
-  };
+  //   if (tel === '') {
+  //     setTelefonoFB({ ...telefonoFB, text: '', color: '' });
+  //     submitBtn.disabled = true;
+  //   } else {
+  //     if (tel.match(pattern) != null && telefono.length >= 6) {
+  //       setTelefonoFB({
+  //         ...telefonoFB,
+  //         text: 'El nummero de telefono es correcto',
+  //         color: '#7CBD1E',
+  //       });
+  //       submitBtn.disabled = false;
+  //     } else {
+  //       setTelefonoFB({
+  //         ...telefonoFB,
+  //         text: 'Solo numeros, minimo 7',
+  //         color: '#CC3636',
+  //       });
+  //       submitBtn.disabled = true;
+  //     }
+  //   }
+  // };
 
-  const handlePickBirth = (e) => {
-    //cambiar fotmato fecha de nacimiento
-    const fechaAdaptada = e.target.value.split('-').join('');
-    setNacimiento(fechaAdaptada);
-  };
+  // const submitAlumnoForm = (e) => {
+  //   e.preventDefault();
+  //   console.log(nacimiento);
+  //   setNombreFB({ ...nombreFB, text: '', color: '' });
+  //   setTelefonoFB({ ...telefonoFB, text: '', color: '' });
+  //   setNacimiento('');
+  //   setAlumnosLoader((prevValue) => !prevValue);
+  //   setActive(false);
+  //   const requestOptions = {
+  //     method: 'POST',
+  //     body: JSON.stringify({
+  //       nombre: nombre,
+  //       telefono: telefono,
+  //       fechanac: nacimiento,
+  //       esalumno: true,
+  //     }),
+  //   };
 
-  const submitAlumnoForm = (e) => {
-    e.preventDefault();
-    console.log(nacimiento);
-    setNombreFB({ ...nombreFB, text: '', color: '' });
-    setTelefonoFB({ ...telefonoFB, text: '', color: '' });
-    setNacimiento('');
-    setAlumnosLoader((prevValue) => !prevValue);
-    setActive(false);
-    const requestOptions = {
-      method: 'POST',
-      body: JSON.stringify({
-        nombre: nombre,
-        telefono: telefono,
-        fechanac: nacimiento,
-        esalumno: true,
-      }),
-    };
-
-    fetch(`${URL_BASE}persona`, requestOptions)
-      .then((response) => response.json())
-      .then((response) => setActAlumnos((v) => !v));
-  };
+  //   fetch(`${URL_BASE}persona`, requestOptions)
+  //     .then((response) => response.json())
+  //     .then((response) => setActAlumnos((v) => !v));
+  // };
 
   const handleCloseForm = () => {
     setActive(false);
@@ -139,10 +137,11 @@ const AgregarAlumno = ({
             x
           </button>
           <h2>Nuevo Alumno</h2>
-          <form action="" id="alumno-add-form" onSubmit={submitAlumnoForm}>
+          <form action="" id="alumno-add-form" onSubmit={handleSubmitAlumnoForm}>
             <div className="inputlabel">
-              <InputComponent
+              <InputReComponent
                 type={'text'}
+                name={'nombre'}
                 className={'alumno-add-form-input'}
                 placeholder={'Nombre'}
                 onChangeFuncion={handleChangeName}
@@ -152,8 +151,9 @@ const AgregarAlumno = ({
               </p>
             </div>
             <div className="inputlabel">
-              <InputComponent
+              <InputReComponent
                 type={'text'}
+                name={'telefono'}
                 id="telefonotinput"
                 className={'alumno-add-form-input'}
                 placeholder={'Telefono'}
@@ -166,7 +166,7 @@ const AgregarAlumno = ({
                 {telefonoFB.text}
               </p>
             </div>
-            <NacimientoComponent setNacimiento={setNacimiento} />
+            <NacimientoComponent alumnoForm={alumnoForm} setNacimiento={setAlumnoForm} />
             <button id="alumno-add-form-addBtn" type="sumbit" disabled>
               <FontAwesomeIcon id="canchas-add-form-btn" icon={faPlusCircle} />
             </button>
